Выполнена разметка элемента списка

+++ app/src/main/java/com/anatoliykichuk/cardiolog/domain/CardioLog.kt
@@ -4,9 +4,9 @@ import java.util.Date
 
 data class CardioLog(
     val date: Date,
-    val diastolicPressure: Double,
-    val systolicPressure: Double,
-    val pulse: Double,
+    val diastolicPressure: Int,
+    val systolicPressure: Int,
+    val pulse: Int,
 ) {
 
 }
